refactor: timedlyrics components move into files
